hack/env.sh: move checking of environment variables outside SKIP_VAR_SET block
When  "$SKIP_VAR_SET" is unset and the environment variables fallback to
the default, the check for valid values should be done. Move the check
out of the $SKIP_VAR_SET block for that.

For the current "hack/env.sh" this maybe not make an actual difference,
because probably the code to assign default values will ensure that
always valid value are set. Note that the openshift variant of the above
code will detect the default via skopeo, which can fail. For that
reason, this change makes more sense for openshift. However, also for
the current code, performing the same error checking after filling out
default values, ensures that the detected values are considered valid
Even if that is in fact always the case, it's not entirely trivial to
see.

Cherry-picked from commit [1] and conflicts resolved.

[1] https://github.com/k8snetworkplumbingwg/sriov-network-operator/commit/ed25caaa819dda402735c82443bcb03be9a9f21a

Signed-off-by: Thomas Haller <thaller@redhat.com>