hack/env.sh: fix detecting default values for images
Note that most of quay.io repositories (not all of them) are depublished
and the detection no longer works:

    FATA[0001] Error parsing image name "docker://quay.io/openshift/origin-sriov-cni": reading manifest latest in quay.io/openshift/origin-sriov-cni: unauthorized: access to the requested resource is not authorized

Together that with [1],  the calling shell scripts now `set -e`, which
exits the script in such errors.

This means, `make deploy-setup` fails unless $SKIP_VAR_SET is set. That
is even if the caller *did* set the missing environment variables for
the images. Note that doing that is not useless, because while several
quay.io repositories no longer exist, others like [2] still do.  So you
can set the ones we must, don't set $SKIP_VAR_SET, and let the others be
detected as in the past.

Fix this by reworking the detection.

The patch does not change URLs like "quay.io/openshift/origin-sriov-cni"
because there is no better default. If the caller doesn't set the
environment variables like SRIOV_CNI_IMAGE, then deploy is going to fail
either way.

[1] 624a16feb327 ('Let "depoly-setup.sh" and "undeploy.sh" fail on error')
[2] https://quay.io/openshift/origin-kube-rbac-proxy

Signed-off-by: Thomas Haller <thaller@redhat.com>