hack/env.sh: fix detecting default values for images
Note that most of quay.io repositories (not all of them) are depublished
and the detection no longer works:

    FATA[0001] Error parsing image name "docker://quay.io/openshift/origin-sriov-cni": reading manifest latest in quay.io/openshift/origin-sriov-cni: unauthorized: access to the requested resource is not authorized

Together that with [1], the calling shell scripts now `set -e`, which
exit the script on such errors.

This means, `make deploy-setup` fails unless $SKIP_VAR_SET is set. That
is even if the caller *did* set the missing environment variables for
the images becase previouls the code would always try to deteminte the
default value (which fails) even if it's not used. Note that not setting
$SKIP_VAR_SET could stil be useful, because while several quay.io
repositories no longer exist, others like [2] still do. So you can set
the ones we must, don't set $SKIP_VAR_SET, and let the others be
detected as previously.

Fix this by reworking the detection.

[1] 624a16feb327 ('Let "depoly-setup.sh" and "undeploy.sh" fail on error')
[2] https://quay.io/openshift/origin-kube-rbac-proxy

Signed-off-by: Thomas Haller <thaller@redhat.com>